# src/main/java/com/event/ticketbooking/auth/model/AuthResponse.java
package com.event.ticketbooking.auth.model;

public class AuthResponse {
    private String accessToken;
    private String tokenType;
    private Long expiresIn;
    private UserDto user;

    public AuthResponse() {
    }

    public AuthResponse(String accessToken, String tokenType, Long expiresIn, UserDto user) {
        this.accessToken = accessToken;
        this.tokenType = tokenType;
        this.expiresIn = expiresIn;
        this.user = user;
    }

    public String getAccessToken() {
        return accessToken;
    }

    public void setAccessToken(String accessToken) {
        this.accessToken = accessToken;
    }

    public String getTokenType() {
        return tokenType;
    }

    public void setTokenType(String tokenType) {
        this.tokenType = tokenType;
    }

    public Long getExpiresIn() {
        return expiresIn;
    }

    public void setExpiresIn(Long expiresIn) {
        this.expiresIn = expiresIn;
    }

    public UserDto getUser() {
        return user;
    }

    public void setUser(UserDto user) {
        this.user = user;
    }
}