# src/main/java/com/event/ticketbooking/auth/model/UserDto.java
package com.event.ticketbooking.auth.model;

public class UserDto {
    private String username;
    private String email;
    private UserRole role;
    private UserStatus status;

    public UserDto() {
    }

    public UserDto(String username, String email, UserRole role, UserStatus status) {
        this.username = username;
        this.email = email;
        this.role = role;
        this.status = status;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public UserRole getRole() {
        return role;
    }

    public void setRole(UserRole role) {
        this.role = role;
    }

    public UserStatus getStatus() {
        return status;
    }

    public void setStatus(UserStatus status) {
        this.status = status;
    }
}