# src/main/java/com/event/ticketbooking/event/model/EventCategory.java
package com.event.ticketbooking.event.model;

public class EventCategory {
    private String name;

    public EventCategory() {
    }

    public EventCategory(String name) {
        this.name = name;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}